integrate cloudinary + change image url (string) option to image upload (file)

+++ contollers/listings.js
@@ -1,3 +1,4 @@
+const { cloudinary } = require("../CloudConfig.js");
 const Listing = require("../models/listing.js");
 
 module.exports.index = async (req, res) => {
@@ -24,8 +25,12 @@ module.exports.showListing = async (req, res) => {
 
 module.exports.createListing = async (req, res, next) => {
     // console.log("RECEIVED BODY:", req.body);
+    let url = req.file.path;
+    let filename = req.file.filename;
+    // console.log(url,"..", filename);
     const newListing = new Listing(req.body.listing);
     newListing.owner = req.user._id;
+    newListing.image = { url, filename };
     await newListing.save();
     req.flash("success", "New Listing Created!");
     res.redirect("/listings");
@@ -38,7 +43,10 @@ module.exports.renderEditForm = async (req, res) => {
         req.flash("error", "Listing you are requesting for does not exist!");
         return res.redirect("/listings");
     }
-    res.render("listings/edit.ejs", { listing });
+
+    let originalImageUrl = listing.image.url;
+    originalImageUrl = originalImageUrl.replace("/upload", "/upload/w_300,h_200,c_fill");
+    res.render("listings/edit.ejs", { listing, originalImageUrl });
 };
 
 module.exports.updateListing = async (req, res) => {
@@ -49,14 +57,43 @@ module.exports.updateListing = async (req, res) => {
         return res.redirect(`/listings/${id}`);
     }
 
-    await Listing.findByIdAndUpdate(id, { ...req.body.listing });
+    let newListing = await Listing.findByIdAndUpdate(id, { ...req.body.listing }, { new: true });
+
+    if (typeof req.file !== "undefined") {
+        // Delete Old Image
+        if (listing.image && listing.image.filename) {
+            await cloudinary.uploader.destroy(listing.image.filename);
+        }
+        // Save New Image
+        let url = req.file.path;
+        let filename = req.file.filename;
+        newListing.image = { url, filename };
+        await newListing.save();
+    }
+
     req.flash("success", "Listing Updated!");
     res.redirect(`/listings/${id}`);
 };
 
 module.exports.deleteListing = async (req, res) => {
     let { id } = req.params;
+
+    // Find listing first (to get image public_id)
+    let listing = await Listing.findById(id);
+
+    if (!listing) {
+        req.flash("error", "Listing not found!");
+        return res.redirect("/listings");
+    }
+
+    // Delete image from Cloudinary
+    if (listing.image && listing.image.filename) {
+        await cloudinary.uploader.destroy(listing.image.filename);
+    }
+
+    // Delete listing from DB
     await Listing.findByIdAndDelete(id);
+
     req.flash("success", "Listing Deleted!");
-    res.redirect("/listings")
+    res.redirect("/listings");
 };
